Simplify text orientation correction
svg.text will now correct orientation of text to always be 90 and 180

Streamline how instance info is passed to Svg so that things are a
little more sensible.

Special-case symbol fields, since that's the only one that actually
needs special-casing (aside from pins which are super complex for other
reasons).